Splitting functionality up properly
